# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "blueeconomy-insurance",
  "version": "0.1.0"
 },
 "paths": {
...
  "/v1/policies/{policy_number}": {
   "get": {
    "tags": [
     "policies"
    ],
    "summary": "Get Policy",
    "operationId": "get_policy_v1_policies__policy_number__get",
    "parameters": [
     {
      "name": "policy_number",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Policy Number"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Policy V1 Policies  Policy Number  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/policies/{policy_number}/credential": {
   "get": {
    "tags": [
     "policies"
    ],
    "summary": "Get Policy Credential",
    "operationId": "get_policy_credential_v1_policies__policy_number__credential_get",
    "parameters": [
     {
      "name": "policy_number",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Policy Number"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Policy Credential V1 Policies  Policy Number  Credential Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/policies/{policy_number}/endorsements": {
   "post": {
    "tags": [
     "policies"
    ],
    "summary": "Add Endorsement",
    "operationId": "add_endorsement_v1_policies__policy_number__endorsements_post",
    "parameters": [
     {
      "name": "policy_number",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Policy Number"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EndorsementIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Add Endorsement V1 Policies  Policy Number  Endorsements Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "policies"
    ],
    "summary": "List Endorsements",
    "operationId": "list_endorsements_v1_policies__policy_number__endorsements_get",
    "parameters": [
     {
      "name": "policy_number",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Policy Number"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Endorsements V1 Policies  Policy Number  Endorsements Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/policies/{policy_number}:cancel": {
   "post": {
    "tags": [
     "policies"
    ],
    "summary": "Cancel Policy",
    "operationId": "cancel_policy_v1_policies__policy_number__cancel_post",
    "parameters": [
     {
      "name": "policy_number",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Policy Number"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CancelIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Cancel Policy V1 Policies  Policy Number  Cancel Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/status-list/{purpose}": {
   "get": {
    "tags": [
     "policies"
    ],
    "summary": "Get Status List",
    "description": "Public, unauthenticated: the signed bitstring status-list credential\nis itself the authenticity object (verifiers check the proof).",
    "operationId": "get_status_list_v1_status_list__purpose__get",
    "parameters": [
     {
      "name": "purpose",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Purpose"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Status List V1 Status List  Purpose  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/v1/integrations/isr/evidence": {
   "post": {
    "tags": [
     "integrations"
    ],
    "summary": "Ingest Isr Evidence",
    "operationId": "ingest_isr_evidence_v1_integrations_isr_evidence_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IsrEvidenceIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Ingest Isr Evidence V1 Integrations Isr Evidence Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/v1/partners/{partner_id}:call": {
   "post": {
    "tags": [
     "integrations"
    ],
    "summary": "Call Partner",
    "operationId": "call_partner_v1_partners__partner_id__call_post",
    "parameters": [
     {
      "name": "partner_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Partner Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PartnerCallIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Call Partner V1 Partners  Partner Id  Call Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CancelIn": {
    "properties": {
     "reason": {
      "type": "string",
      "maxLength": 2000,
      "title": "Reason",
      "default": ""
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "CancelIn"
   },
...
   "EndorsementIn": {
    "properties": {
     "kind": {
      "type": "string",
      "enum": [
       "EXTENSION",
       "VALUE_CHANGE",
       "ASSURED_CHANGE",
       "CANCELLATION",
       "REINSTATEMENT"
      ],
      "title": "Kind"
     },
     "premium_delta_kobo": {
      "type": "integer",
      "title": "Premium Delta Kobo",
      "default": 0
     },
     "detail": {
      "additionalProperties": true,
      "type": "object",
      "title": "Detail"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "kind"
    ],
    "title": "EndorsementIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "IsrEvidenceIn": {
    "properties": {
     "envelope": {
      "additionalProperties": true,
      "type": "object",
      "title": "Envelope"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "envelope"
    ],
    "title": "IsrEvidenceIn"
   },
...
   "PartnerCallIn": {
    "properties": {
     "operation": {
      "type": "string",
      "enum": [
       "quote-cede-offer",
       "claim-notification",
       "policy-evidence-pack"
      ],
      "title": "Operation"
     },
     "payload": {
      "additionalProperties": true,
      "type": "object",
      "title": "Payload"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "operation",
     "payload"
    ],
    "title": "PartnerCallIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}